requirements.txt
flask==2.0.1
flask-cors==3.0.10
pymongo==4.6.1
python-dotenv==1.0.0
werkzeug==2.0.1
requests~=2.32.3